+++ src/main/java/domain/menu/Amount.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.menu;
 
 import java.util.Objects;
