+++ src/main/java/domain/price/Payment.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.price;
 
 import java.util.Arrays;
