+++ src/main/java/domain/table/TableRepository.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.table;
 
 import java.util.ArrayList;
 import java.util.Collections;
