+++ src/test/java/domain/menu/AmountTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.menu;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +11,7 @@ class AmountTest {
     @Test
     @DisplayName("Amount 생성 테스트")
     void construct() {
-        assertThat(new Amount(10)).isNotNull();
+        Assertions.assertThat(new Amount(10)).isNotNull();
     }
 
     @Test
