+++ src/test/java/domain/price/PriceTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.price;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
