+++ src/test/java/domain/table/TableTest.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.table;
 
+import domain.menu.MenuRepository;
+import domain.table.Table;
+import domain.table.TableRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
